djangotutorial/pyproject.toml
[project]
name = "placeholder-api"
version = "0.0.1"
requires-python = ">= 3.11"
dependencies = [
  "Django>=6.0.3",
  "djangorestframework==3.16.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.15.10", "basedpyright>=1.39.0", "coverage>=7.13.5"]
test = ["jsonschema>=4.26.0"]

[tool.basedpyright]
typeCheckingMode = "standard"
